Phase 4: pull the three catalog data loads into hooks
Three independent fetch/state slices sat inline in ArcadeCatalog, each
with its own cancellation and readiness flag, interleaved with everything
else the component does:

- useCatalogSortSignals  the viewer-scoped signals cache, the parallel
  fetch, and payloadToSignals/sameCatalogSortSignals/initialSignals, the
  three pure helpers only that fetch used
- useCreatorShelf        creator items, the auth-gated reset and the
  30s poll, returning the mySlugs set the grid actually consumes
- useFeaturedPool        the curated pool and its timeout-guarded fetch,
  with FEATURED_POOL_TIMEOUT_MS

Each returns only what the surface reads: creatorItems never leaves
useCreatorShelf, since mySlugs was its single consumer.

ArcadeCatalog 678 -> 545. No behaviour change: same effects, same deps,
same order of the state they set.
